Scrapy-SchoolMarket/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

DBUtils==3.1.0
itemadapter==0.9.0
jieba==0.42.1
matplotlib==3.9.0
mysql-connector-python==8.3.0
numpy==1.26.4
paddleocr==2.7.3
pandas==2.2.2
pillow==10.3.0
requests==2.31.0
Scrapy==2.11.1
scipy==1.12.0
torch==2.3.1
transformers==4.42.3
